contract-analyzer/requirements.txt
streamlit
pytesseract
pdf2image
Pillow
spacy
transformers
torch
